[#7] feat :rocket: PlayViewController로 넘어갈 때 gameId, gameName 가지고 네트워크 요청
