+++ iOS/BaseballGame/BaseballGame/Utilities/Network/UseCase/GameRoomInfoUseCase.swift
@@ -1,5 +1,5 @@
 //
-//  MockGameRoomInfoUseCase.swift
+//  GameRoomInfoUseCase.swift
 //  BaseballGame
 //
 //  Created by delma on 2020/05/13.
@@ -7,25 +7,6 @@
 //
 
 import Foundation
-
-struct MockGameRoomInfoUseCase {
-      private let task = NetworkTask(dispatcher: NetworkDispatcher())
-      
-      class MockGameRoomRequest: Request {
-        var path: String = EndPoints.fakeGameRoomEmpty
-      }
-      
-      func requestGameRoomInfoStub(handler: @escaping (Any) -> Void) {
-          task.perform(request: MockGameRoomRequest(), dataType: GameRoomEmpty.self) { result in
-              switch result {
-              case .success(let decodedData):
-                  handler(decodedData)
-              case .failure(let error):
-                  NotificationCenter.default.post(name: .networkError, object: nil, userInfo: ["error":error])
-              }
-          }
-      }
-}
 
 struct GameRoomInfoUseCase {
       private let task = NetworkTask(dispatcher: NetworkDispatcher())
